pyproject.toml
[tool.poetry]
name = "ska-tango-examples"
version = "0.4.32"
description = "SKA Tango Examples"
authors = ["Matteo Di Carlo <tmatteo.dicarlo@inaf.it>"]
license = "BSD-3-Clause"
packages = [
    { include = "ska_tango_examples", from = "src" }
]

[tool.poetry.scripts]
TangoExampleCounterDS = 'ska_tango_examples.counter.Counter:main'
TangoExampleMotorDS = 'ska_tango_examples.basic_example.Motor:main'
TangoExampleEventReceiverDS = 'ska_tango_examples.basic_example.EventReceiver:main'
TangoExampleTabataDS = 'ska_tango_examples.tabata.Tabata:main'
TangoExampleAsyncTabataDS = 'ska_tango_examples.tabata.AsyncTabata:main'

[[tool.poetry.source]]
name = 'ska-nexus'
url = 'https://artefact.skao.int/repository/pypi-internal/simple'

[[tool.poetry.source]]
name = "PyPI"

[tool.poetry.dependencies]
python = ">=3.10,<=3.12"
debugpy = "^1.5.1"
numpy = "^1.26.1"
pytango = "^9.5.0"
ska-k8s-config-exporter = "^0.0.4"
ska-ser-log-transactions = "*"
ska-tango-base = "^0.20.2"
notebook = "^7.0.6"
jupyterlab = "^4.0.8"
jupyterhub = "^4.0.2"
jupyter-server = "^2.10.0"
fastjsonschema = "^2.19.1"
attrs = "^23.2.0"
referencing = "^0.35.1"
rpds-py = "^0.18.0"
jsonschema-specifications = "^2023.12.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
coverage = "^7.5.1"
flake8 = "^7.0.0"
isort = "^5.13.0"
nbmake = {extras = ["all"], version = "^1.5.0"}
nbqa = {extras = ["all"], version = "^1.8.0"}
pylint = "^3.1.0"
pylint-junit = "^0.3.2"
pytest = "^8.2.0"
pytest-bdd = "^7.1.2"
pytest-cov = "^5.0.0"
pytest-forked = "^1.6.0"
pytest-json-report = "^1.5.0"
pytest-mock = "^3.14.0"
pytest-repeat = "^0.9.3"
pytest-timeout = "^2.3.1"
yamllint = "^1.32.0"
ska-tango-testing = "^0.6.1"

[tool.poetry.group.docs.dependencies]
myst-parser = "^3.0.0"
ska-ser-sphinx-theme = "^0.1.2"
Sphinx = "^7.3.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 79

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
